Creating ConnectionHandler - Create ConnectionHandler.cpp - Create ConnectionHandler.h - Moving socket connection logic to ConnectionHandler
